@@ -11,42 +11,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~8.0.1",
-    "@angular/common": "~8.0.1",
-    "@angular/compiler": "~8.0.1",
-    "@angular/core": "~8.0.1",
-    "@angular/forms": "~8.0.1",
-    "@angular/platform-browser": "~8.0.1",
-    "@angular/platform-browser-dynamic": "~8.0.1",
-    "@angular/router": "~8.0.1",
-    "@ngrx/data": "^8.0.1",
-    "@ngrx/effects": "^8.0.1",
-    "@ngrx/entity": "^8.0.1",
-    "@ngrx/store": "^8.0.1",
-    "@ngrx/store-devtools": "^8.0.1",
-    "rxjs": "~6.4.0",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.9.1"
+    "@angular/animations": "^10.0.8",
+    "@angular/common": "^10.0.8",
+    "@angular/compiler": "^10.0.8",
+    "@angular/core": "^10.0.8",
+    "@angular/forms": "^10.0.8",
+    "@angular/platform-browser": "^10.0.8",
+    "@angular/platform-browser-dynamic": "^10.0.8",
+    "@angular/router": "^10.0.8",
+    "@ngrx/data": "^10.0.0",
+    "@ngrx/effects": "^10.0.0",
+    "@ngrx/entity": "^10.0.0",
+    "@ngrx/router-store": "^10.0.0",
+    "@ngrx/store": "^10.0.0",
+    "@ngrx/store-devtools": "^10.0.0",
+    "ngrx-undo": "^8.0.0",
+    "rxjs": "^6.6.2",
+    "tslib": "^2.0.1",
+    "zone.js": "^0.10.3"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "~0.800.0",
-    "@angular/cli": "~8.0.3",
-    "@angular/compiler-cli": "~8.0.1",
-    "@angular/language-service": "~8.0.1",
-    "@types/node": "~8.9.4",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.3",
-    "codelyzer": "^5.0.0",
-    "jasmine-core": "~3.4.0",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "~4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~2.0.1",
-    "karma-jasmine-html-reporter": "^1.4.0",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.15.0",
-    "typescript": "~3.4.3"
+    "@angular-devkit/build-angular": "^0.1000.5",
+    "@angular/cli": "^10.0.5",
+    "@angular/compiler-cli": "^10.0.8",
+    "@angular/language-service": "^10.0.8",
+    "@types/jasmine": "^3.5.12",
+    "@types/jasminewd2": "^2.0.8",
+    "@types/node": "^14.0.27",
+    "codelyzer": "^6.0.0",
+    "eslint": "^7.6.0",
+    "jasmine-core": "^3.6.0",
+    "jasmine-spec-reporter": "^5.0.2",
+    "karma": "^5.1.1",
+    "karma-chrome-launcher": "^3.1.0",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-jasmine": "^3.3.1",
+    "karma-jasmine-html-reporter": "^1.5.4",
+    "protractor": "^7.0.0",
+    "ts-node": "^8.10.2",
+    "tslint": "^6.1.3",
+    "typescript": "~3.9.7"
   }
 }
