@@ -3,10 +3,10 @@
 // The list of file replacements can be found in `angular.json`.
 
 export const environment = {
-  production: false,
-  appApi: {
-    baseUrl: 'http://localhost:8080'
-  }
+    production: false,
+    appApi: {
+        baseUrl: 'http://localhost:3000'
+    }
 };
 
 /*
